Move setupCompute and related functions out of GraphModel and into PlottedFunctionAdornmentModel.
